pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "gangzi-trading"
version = "0.1.0"
description = "钢子出击 - 加密货币自动交易系统"
requires-python = ">=3.11"
license = "MIT"

dependencies = [
    # Web 框架
    "fastapi>=0.129.0",
    "uvicorn[standard]>=0.34.3",

    # 数据库
    "sqlalchemy[asyncio]>=2.0.46",
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",

    # 认证
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=5.0.0",

    # HTTP 客户端
    "httpx>=0.28.1",

    # WebSocket
    "websockets>=15.0.1",

    # 数据处理
    "pandas>=3.0.0",
    "pandas-ta>=0.4.71b0",

    # 配置管理
    "python-dotenv>=1.2.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",

    # 定时任务
    "apscheduler>=3.11.2",

    # 表单解析
    "python-multipart>=0.0.22",

    # 加密（API Key 加密存储）
    "cryptography>=46.0.5",

    # 交易所 API（自动交易）
    "ccxt>=4.5.38",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.9.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F"]
ignore = ["E501", "W293", "W291"]

[tool.setuptools.packages.find]
include = ["backend*"]
